+++ src/Access/Ilias.php
@@ -3,8 +3,6 @@
 namespace srag\Plugins\H5P\Access;
 
 use ilH5PPlugin;
-use ilWACPath;
-use ilWACSecurePath;
 use srag\DIC\H5P\DICTrait;
 use srag\Plugins\H5P\Utils\H5PTrait;
 
@@ -15,8 +13,7 @@ use srag\Plugins\H5P\Utils\H5PTrait;
  *
  * @author  studer + raimann ag - Team Custom 1 <support-custom1@studer-raimann.ch>
  */
-final class Ilias/* implements ilWACCheckingClass*/
-{
+final class Ilias {
 
 	use DICTrait;
 	use H5PTrait;
@@ -45,33 +42,4 @@ final class Ilias/* implements ilWACCheckingClass*/
 	public function __construct() {
 
 	}
-	/**
-	 *
-	 * /
-	 * public function registerWAC() {
-	 * //ilWACSignedPath::signFolderOfStartFile(self::h5p()->getH5PFolder() . "/dummy.js");
-	 *
-	 * /**
-	 * @var ilWACSecurePath $path
-	 * /
-	 * $path = ilWACSecurePath::findOrGetInstance("h5p");
-	 * $path->setPath("h5p");
-	 *
-	 * $path->setCheckingClass(self::class);
-	 * $path->setInSecFolder(false);
-	 * $path->setComponentDirectory(__DIR__ . "/../../");
-	 *
-	 * $path->store();
-	 * }
-	 *
-	 *
-	 * /**
-	 *
-	 * @param ilWACPath     $ilWACPath
-	 *
-	 * @return bool
-	 * /
-	 * public function canBeDelivered(ilWACPath $ilWACPath) {
-	 * return true;
-	 * }*/
 }