+++ src/Content/Content.php
@@ -22,6 +22,8 @@ class Content extends ActiveRecord {
 	use H5PTrait;
 	const TABLE_NAME = "rep_robj_xhfp_cont";
 	const PLUGIN_CLASS_NAME = ilH5PPlugin::class;
+	const PARENT_TYPE_OBJECT = "object";
+	const PARENT_TYPE_PAGE = "page";
 
 
 	/**
@@ -131,7 +133,7 @@ class Content extends ActiveRecord {
 	 *
 	 * @return Content[]
 	 */
-	public static function getContentsByObject($obj_id, $parent_type = "object") {
+	public static function getContentsByObject($obj_id, $parent_type = self::PARENT_TYPE_OBJECT) {
 		/**
 		 * @var Content[] $h5p_contents
 		 */
@@ -156,7 +158,7 @@ class Content extends ActiveRecord {
 	 *
 	 * @return array
 	 */
-	public static function getContentsByObjectArray($obj_id, $parent_type = "object") {
+	public static function getContentsByObjectArray($obj_id, $parent_type = self::PARENT_TYPE_OBJECT) {
 		$h5p_contents = self::where([
 			"obj_id" => $obj_id,
 			"parent_type" => $parent_type
@@ -389,7 +391,7 @@ class Content extends ActiveRecord {
 	 * @con_fieldtype    text
 	 * @con_is_notnull   true
 	 */
-	protected $parent_type = "object";
+	protected $parent_type = self::PARENT_TYPE_OBJECT;
 	/**
 	 * @var int
 	 *
@@ -399,6 +401,14 @@ class Content extends ActiveRecord {
 	 * @con_is_notnull   true
 	 */
 	protected $sort;
+	/**
+	 * @var string[]
+	 *
+	 * @con_has_field    true
+	 * @con_fieldtype    text
+	 * @con_is_notnull   true
+	 */
+	protected $uploaded_files = [];
 
 
 	/**
@@ -424,7 +434,9 @@ class Content extends ActiveRecord {
 			case "created_at":
 			case "updated_at":
 				return self::h5p()->timestampToDbDate($field_value);
-				break;
+
+			case "uploaded_files":
+				return json_encode($field_value);
 
 			default:
 				return NULL;
@@ -446,12 +458,10 @@ class Content extends ActiveRecord {
 			case "disable":
 			case "sort":
 				return intval($field_value);
-				break;
 
 			case "created_at":
 			case "updated_at":
 				return self::h5p()->dbDateToTimestamp($field_value);
-				break;
 
 			case "obj_id":
 				if ($field_value !== NULL) {
@@ -459,7 +469,9 @@ class Content extends ActiveRecord {
 				} else {
 					return NULL;
 				}
-				break;
+
+			case "uploaded_files":
+				return json_decode($field_value);
 
 			default:
 				return NULL;
@@ -806,5 +818,21 @@ class Content extends ActiveRecord {
 	 */
 	public function setSort($sort) {
 		$this->sort = $sort;
+	}
+
+
+	/**
+	 * @return array
+	 */
+	public function getUploadedFiles() {
+		return $this->uploaded_files;
+	}
+
+
+	/**
+	 * @param string[] $uploaded_files
+	 */
+	public function setUploadedFiles(array $uploaded_files) {
+		$this->uploaded_files = $uploaded_files;
 	}
 }