+++ src/Content/ShowContent.php
@@ -312,6 +312,8 @@ class ShowContent {
 
 		$h5p_tpl->setVariable("H5P_CONTENT_ID", $content_id);
 
+		$h5p_tpl->parseCurrentBlock();
+
 		return self::output()->getHTML($h5p_tpl);
 	}
 
@@ -326,7 +328,7 @@ class ShowContent {
 	 */
 	public function setFinished($content_id, $score, $max_score, $opened, $finished, $time = NULL) {
 		$h5p_content = Content::getContentById($content_id);
-		if ($h5p_content !== NULL && $h5p_content->getParentType() === "object") {
+		if ($h5p_content !== NULL && $h5p_content->getParentType() === Content::PARENT_TYPE_OBJECT) {
 			$object = H5PObject::getObjectById($h5p_content->getObjId());
 		} else {
 			$object = NULL;
@@ -346,7 +348,7 @@ class ShowContent {
 		} else {
 			// Prevent update result on a repository object with "Solve only once"
 			if ($object !== NULL && $object->isSolveOnlyOnce()) {
-				die();
+				return;
 			}
 		}
 
@@ -387,7 +389,7 @@ class ShowContent {
 	 */
 	public function contentsUserData($content_id, $data_id, $sub_content_id, $data = NULL, $preload = false, $invalidate = false) {
 		$h5p_content = Content::getContentById($content_id);
-		if ($h5p_content !== NULL && $h5p_content->getParentType() === "object") {
+		if ($h5p_content !== NULL && $h5p_content->getParentType() === Content::PARENT_TYPE_OBJECT) {
 			$object = H5PObject::getObjectById($h5p_content->getObjId());
 		} else {
 			$object = NULL;
@@ -417,7 +419,7 @@ class ShowContent {
 				} else {
 					// Prevent update user data on a repository object with "Solve only once". But some contents may store date with editor so check has results
 					if ($object !== NULL && $object->isSolveOnlyOnce() && Result::hasContentResults($h5p_content->getContentId())) {
-						die();
+						return NULL;
 					}
 				}